+++ mkparts/mkparts.c
@@ -25,8 +25,8 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-#include "matroska/matroska.h"
-#include "matroska/matroska_sem.h"
+#include "matroska2/matroska.h"
+#include "matroska2/matroska_sem.h"
 #include "mkparts_project.h"
 
 static textwriter *StdErr = NULL;
