+++ spectool/schema_2_matroska_sem_c.xsl
@@ -39,9 +39,9 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-#include "matroska/matroska.h"
-#include "matroska/matroska_sem.h"
-#include "matroska/matroska_internal.h"
+#include "matroska2/matroska.h"
+#include "matroska2/matroska_sem.h"
+#include "matroska2/matroska_internal.h"
 
 <xsl:call-template name="output-element-declaration">
     <xsl:with-param name="node" select="ebml:element[@path='\Segment']"/>
