+++ spectool/schema_2_matroska_sem_h.xsl
@@ -39,7 +39,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-#include "matroska/matroska.h"
+#include "matroska2/matroska.h"
 
 #ifndef MATROSKA_SEMANTIC_H
 #define MATROSKA_SEMANTIC_H
